feat: reformat inputs module
